+++ pyproject.toml
@@ -3,7 +3,7 @@ name = "pivot"
 version = "0.1.0-dev"
 description = "High-performance Python pipeline tool with automatic code change detection"
 readme = "README.md"
-requires-python = ">=3.13,<3.14"
+requires-python = ">=3.13"
 license = { text = "TBD" }
 authors = [{ name = "Internal Team" }]
 dependencies = [
@@ -35,6 +35,7 @@ dev = [
     "pytest-cov>=4.0",
     "pytest-mock>=3.15.1",
     "pytest-watcher>=0.6.2",
+    "pytest-xdist>=3.8.0",
     "pytest>=7.0",
     "ruff>=0.8.0",
 ]
