+++ tests/cli/test_console.py
