+++ tests/compat/test_dvc_compat_integration.py
