+++ tests/config/test_validation.py
@@ -8,6 +8,7 @@ import pytest
 
 from pivot import registry
 from pivot.exceptions import ValidationError
+from pivot.registry import ValidationMode
 
 
 def test_duplicate_stage_name_raises_error() -> None:
@@ -20,15 +21,15 @@ def test_duplicate_stage_name_raises_error() -> None:
     def stage2() -> None:
         pass
 
-    reg.register(stage1, name="process", deps=list[str](), outs=list[str]())
+    reg.register(stage1, name="process", deps=list[str](), outs=list[str](), params=None)
 
     with pytest.raises(ValidationError, match="already registered"):
-        reg.register(stage2, name="process", deps=list[str](), outs=list[str]())
+        reg.register(stage2, name="process", deps=list[str](), outs=list[str](), params=None)
 
 
 def test_duplicate_stage_name_with_warning_mode() -> None:
     """Should log warning but allow registration in WARN mode."""
-    reg = registry.StageRegistry(validation_mode=registry.ValidationMode.WARN)
+    reg = registry.StageRegistry(validation_mode=ValidationMode.WARN)
 
     def stage1() -> None:
         pass
@@ -36,10 +37,10 @@ def test_duplicate_stage_name_with_warning_mode() -> None:
     def stage2() -> None:
         pass
 
-    reg.register(stage1, name="process", deps=list[str](), outs=list[str]())
+    reg.register(stage1, name="process", deps=list[str](), outs=list[str](), params=None)
 
     # Should not raise, just warn
-    reg.register(stage2, name="process", deps=list[str](), outs=list[str]())
+    reg.register(stage2, name="process", deps=list[str](), outs=list[str](), params=None)
 
     # Second registration should overwrite
     assert reg.get("process")["func"] is stage2
@@ -58,6 +59,7 @@ def test_invalid_dep_path_with_parent_traversal() -> None:
             name="process",
             deps=["../secrets/passwords.txt"],
             outs=list[str](),
+            params=None,
         )
 
 
@@ -74,6 +76,7 @@ def test_invalid_out_path_with_parent_traversal() -> None:
             name="process",
             deps=list[str](),
             outs=["../system/file.txt"],
+            params=None,
         )
 
 
@@ -90,6 +93,7 @@ def test_invalid_path_with_null_byte() -> None:
             name="process",
             deps=["data\x00.csv"],
             outs=list[str](),
+            params=None,
         )
 
 
@@ -106,6 +110,7 @@ def test_invalid_path_with_newline() -> None:
             name="process",
             deps=["file\nname.csv"],
             outs=list[str](),
+            params=None,
         )
 
 
@@ -119,15 +124,15 @@ def test_output_conflict_raises_error() -> None:
     def stage2() -> None:
         pass
 
-    reg.register(stage1, name="process1", deps=list[str](), outs=["output.txt"])
+    reg.register(stage1, name="process1", deps=list[str](), outs=["output.txt"], params=None)
 
     with pytest.raises(ValidationError, match="produced by both"):
-        reg.register(stage2, name="process2", deps=list[str](), outs=["output.txt"])
+        reg.register(stage2, name="process2", deps=list[str](), outs=["output.txt"], params=None)
 
 
 def test_output_conflict_with_warning_mode() -> None:
     """Should log warning but allow registration in WARN mode."""
-    reg = registry.StageRegistry(validation_mode=registry.ValidationMode.WARN)
+    reg = registry.StageRegistry(validation_mode=ValidationMode.WARN)
 
     def stage1() -> None:
         pass
@@ -135,10 +140,10 @@ def test_output_conflict_with_warning_mode() -> None:
     def stage2() -> None:
         pass
 
-    reg.register(stage1, name="process1", deps=list[str](), outs=["output.txt"])
+    reg.register(stage1, name="process1", deps=list[str](), outs=["output.txt"], params=None)
 
     # Should not raise, just warn
-    reg.register(stage2, name="process2", deps=list[str](), outs=["output.txt"])
+    reg.register(stage2, name="process2", deps=list[str](), outs=["output.txt"], params=None)
 
 
 def test_empty_stage_name_raises_error() -> None:
@@ -149,7 +154,7 @@ def test_empty_stage_name_raises_error() -> None:
         pass
 
     with pytest.raises(ValidationError, match="cannot be empty"):
-        reg.register(stage1, name="", deps=list[str](), outs=list[str]())
+        reg.register(stage1, name="", deps=list[str](), outs=list[str](), params=None)
 
 
 def test_whitespace_only_stage_name_raises_error() -> None:
@@ -160,7 +165,7 @@ def test_whitespace_only_stage_name_raises_error() -> None:
         pass
 
     with pytest.raises(ValidationError, match="cannot be empty"):
-        reg.register(stage1, name="   ", deps=list[str](), outs=list[str]())
+        reg.register(stage1, name="   ", deps=list[str](), outs=list[str](), params=None)
 
 
 def test_valid_inputs_pass_silently() -> None:
@@ -176,6 +181,7 @@ def test_valid_inputs_pass_silently() -> None:
         name="process",
         deps=["data/input.csv"],
         outs=["data/output.csv"],
+        params=None,
     )
 
     assert "process" in reg.list_stages()
@@ -183,7 +189,7 @@ def test_valid_inputs_pass_silently() -> None:
 
 def test_invalid_path_with_parent_traversal_warn_mode() -> None:
     """Should warn but not raise for path with '..' in WARN mode."""
-    reg = registry.StageRegistry(validation_mode=registry.ValidationMode.WARN)
+    reg = registry.StageRegistry(validation_mode=ValidationMode.WARN)
 
     def stage1() -> None:
         pass
@@ -194,6 +200,7 @@ def test_invalid_path_with_parent_traversal_warn_mode() -> None:
         name="process",
         deps=["../external/data.csv"],
         outs=list[str](),
+        params=None,
     )
 
     assert "process" in reg.list_stages()
@@ -201,12 +208,12 @@ def test_invalid_path_with_parent_traversal_warn_mode() -> None:
 
 def test_invalid_path_with_null_byte_warn_mode() -> None:
     """Should warn but not raise for path with null byte in WARN mode."""
-    reg = registry.StageRegistry(validation_mode=registry.ValidationMode.WARN)
+    reg = registry.StageRegistry(validation_mode=ValidationMode.WARN)
 
     def stage1() -> None:
         pass
 
     # Should not raise, just warn
-    reg.register(stage1, name="process", deps=["bad\x00file.csv"], outs=list[str]())
+    reg.register(stage1, name="process", deps=["bad\x00file.csv"], outs=list[str](), params=None)
 
     assert "process" in reg.list_stages()
