+++ tests/core/test_transitive_deps.py
